Rename LLVM generator modules
